requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==1.10.8
pynput==1.8.1
pyperclip==1.9.0
Pillow==10.2.0
python-dotenv==1.0.1
langchain==0.1.7
langchain-community==0.0.19
langchain-ollama==0.0.3
ollama==0.1.6
websockets==12.0
pyinstaller==6.3.0